+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r124-release...b7fca09299330900f6b7328dbbf9caaf73e30537)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r125-release...bbeb473c942f4d33db70d4bf05f6ebe86b603137)
 
-- Update locales
-- Update patron
+- Add "Show group number" to NameText
+- Make spotlight menu always on-screen
+- Update defensives for Retail
+- Update externals for WotLK
+- Fix icon duration text
+- Update RaidRosterFrame
+- Update ReadyAndPull